+++ public/scripts/movim_utils.js
@@ -144,6 +144,21 @@ var MovimUtils = {
 
         return {'page': page, 'params': str, 'hash': window.location.hash.substr(1)};
     },
+    humanFileSize : function(bytes, si) {
+        var thresh = si ? 1000 : 1024;
+        if(Math.abs(bytes) < thresh) {
+            return bytes + ' B';
+        }
+        var units = si
+            ? ['kB','MB','GB','TB','PB','EB','ZB','YB']
+            : ['KiB','MiB','GiB','TiB','PiB','EiB','ZiB','YiB'];
+        var u = -1;
+        do {
+            bytes /= thresh;
+            ++u;
+        } while(Math.abs(bytes) >= thresh && u < units.length - 1);
+        return bytes.toFixed(1)+' '+units[u];
+    },
     getOrientation : function(file, callback) {
         var reader = new FileReader();